requirements.txt
streamlit
python-dotenv
google-api-python-client
nltk